+++ src/components/WeatherBadge/type.ts
@@ -1,0 +1,16 @@
+type Weather = {
+  sky: string;
+  temperature: string;
+  dust: string;
+};
+
+type WeatherBadge = {
+  weather?: Weather;
+  isRejected: boolean;
+};
+
+export type ContainerProps = {
+  isLoading?: boolean;
+};
+
+export type { WeatherBadge };
